@@ -1,8 +1,10 @@
 module github.com/lstoll/sa2oidc
 
-go 1.14
+go 1.22
 
 require (
 	github.com/pardot/oidc v0.0.0-20200506190725-cbeb86b3b56f
 	gopkg.in/square/go-jose.v2 v2.4.0
 )
+
+require golang.org/x/crypto v0.0.0-20190923035154-9ee001bba392 // indirect
